Split classes into separate files
CHANGELOG_BEGIN
CHANGELOG_END
